+++ song-service/src/main/kotlin/com/epam/service/SongService.kt
@@ -1,9 +1,9 @@
 package com.epam.service
 
-import com.epam.exception.CustomException
+import CustomException
+import com.epam.dto.SongView
 import com.epam.model.Song
 import com.epam.repository.SongRepository
-import com.epam.view.SongView
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 import kotlin.jvm.optionals.getOrNull
@@ -14,7 +14,7 @@ class SongService(
 ) {
 
     @Transactional
-    fun saveSong(songView: SongView): Int {
+    fun saveSong(songView: SongView): Long {
         val song = Song()
 
         with(song) {
@@ -29,11 +29,11 @@ class SongService(
         return songRepository.save(song).id ?: throw CustomException("smth went wrong", 500)
     }
 
-    fun getSong(id: Int): Song =
+    fun getSong(id: Long): Song =
         songRepository.findById(id).getOrNull() ?: throw CustomException("song with id $id wasnt found", 404)
 
     @Transactional
-    fun deleteSong(id: Int): Int {
+    fun deleteSong(id: Long): Long {
         songRepository.deleteById(id)
 
         return id
